Fix email auth
