+++ packages/kit/src/components/ConnectButton/index.ts
@@ -1,0 +1,1 @@
+export { ConnectButton, GoogleWaasConnectButton, AppleWaasConnectButton, EmailConnectButton } from './ConnectButton'
